refactor: split capability negotiation from AppStore (MIM-30)
